requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.10.3
tabulate==0.9.0

apache-airflow-providers-amazon
pandas
boto3
pyspark

pytest==7.4.0
pytest-mock==3.11.1
moto==4.1.11  # for mocking AWS services

python-dotenv==1.0.0

Sphinx==7.1.2
sphinx-rtd-theme==1.3.0


